// mbi-server/proto/user/User.ts
// Original file: ../proto/user/user.proto


export interface User {
  'id'?: (number);
  'firstName'?: (string);
  'lastName'?: (string);
  'companyName'?: (string);
  'email'?: (string);
  'password'?: (string);
}

export interface User__Output {
  'id'?: (number);
  'firstName'?: (string);
  'lastName'?: (string);
  'companyName'?: (string);
  'email'?: (string);
  'password'?: (string);
}